reafactor: split blocks
