+++ packages/blocky-core/src/block/basic.ts
@@ -1,5 +1,4 @@
 import { type IDisposable, type Position } from "blocky-common/es";
-import { elem } from "blocky-common/es/dom";
 import { type HTMLConverter } from "@pkg/helper/htmlConverter";
 import {
   type DataBaseNode,
@@ -8,8 +7,7 @@ import {
   CursorState,
   Changeset,
 } from "@pkg/data";
-import Delta from "quill-delta-es";
-import { Observable, Subject, fromEvent, takeUntil } from "rxjs";
+import { Subject } from "rxjs";
 import { type Editor } from "@pkg/view/editor";
 import { type EditorController } from "@pkg/view/controller";
 import { RenderOption } from "@pkg/view/renderer";
@@ -227,149 +225,4 @@ export class Block implements IDisposable {
   }
 }
 
-/**
- * Blocks except "title"
- *
- * Supports:
- * - copy & paste selection.
- * - drag & drop
- */
-export class ContentBlock extends Block {
-  static DragOverClassName = "blocky-drag-over";
-
-  container: HTMLElement | null = null;
-  contentContainer: HTMLElement | null = null;
-
-  readonly #dragOver = new Subject<DragEvent>();
-  readonly #drop = new Subject<DragEvent>();
-
-  #dragOverBar: HTMLElement | undefined;
-
-  blockDidMount(e: BlockDidMountEvent): void {
-    this.container = e.element;
-  }
-
-  get dragOver$(): Observable<DragEvent> {
-    return this.#dragOver.pipe(takeUntil(this.dispose$));
-  }
-
-  get drop$(): Observable<DragEvent> {
-    return this.#drop.pipe(takeUntil(this.dispose$));
-  }
-
-  setDragOverState(state: BlockDragOverState): void {
-    const container = this.container;
-    if (!container) {
-      return;
-    }
-    switch (state) {
-      case BlockDragOverState.None:
-        container.classList.remove(ContentBlock.DragOverClassName);
-        if (this.#dragOverBar) {
-          this.#dragOverBar.remove();
-          this.#dragOverBar = undefined;
-        }
-        break;
-      case BlockDragOverState.Top:
-      case BlockDragOverState.Bottom: {
-        container.classList.add(ContentBlock.DragOverClassName);
-        const bar = this.createDragOverBar(state === BlockDragOverState.Top);
-        this.#dragOverBar = bar;
-        if (state === BlockDragOverState.Top) {
-          container.prepend(bar);
-        } else {
-          container.append(bar);
-        }
-        break;
-      }
-    }
-  }
-
-  protected initBlockDnd(contentContainer: HTMLElement) {
-    fromEvent<DragEvent>(contentContainer, "dragover").subscribe(
-      this.#dragOver
-    );
-    fromEvent<DragEvent>(contentContainer, "drop").subscribe(this.#drop);
-  }
-
-  protected createDragOverBar(isTop: boolean): HTMLElement {
-    const result = elem("div", "blocky-drag-over-bar");
-    if (isTop) {
-      result.style.top = "0px";
-    } else {
-      result.style.bottom = "0px";
-    }
-    return result;
-  }
-}
-
 export const zeroWidthChar = "\u200b";
-
-/**
- * Base class for the block with a content container
- * Handle the copy & paste selection.
- *
- * If you want to write a custom block, extend this class
- */
-export class CustomBlock extends ContentBlock {
-  #selectSpan: HTMLSpanElement | undefined;
-
-  override blockDidMount(e: BlockDidMountEvent): void {
-    super.blockDidMount(e);
-    const { element, blockDef } = e;
-    const contentContainer = elem("div", "blocky-content");
-    this.contentContainer = contentContainer;
-
-    if (!blockDef.Editable) {
-      contentContainer.contentEditable = "false";
-    }
-
-    element.appendChild(contentContainer);
-    const nonWidthChar = document.createTextNode(zeroWidthChar);
-    this.#selectSpan = elem("span", "blocky-select-span");
-    this.#selectSpan.setAttribute("data-id", e.blockElement.id);
-    this.#selectSpan.appendChild(nonWidthChar);
-    element.append(this.#selectSpan);
-
-    this.initBlockDnd(contentContainer);
-  }
-
-  /**
-   * Select the text to simulate the selection.
-   */
-  override blockFocused(e: BlockFocusedEvent): void {
-    const { selection } = e;
-    const range = document.createRange();
-    range.selectNode(this.#selectSpan!);
-    selection.removeAllRanges();
-    selection.addRange(range);
-  }
-
-  /**
-   * If user input something in the select span.
-   * Convert it to new block.
-   */
-  override blockContentChanged(e: BlockContentChangedEvent): void {
-    const { changeset } = e;
-    if (this.#selectSpan!.textContent !== zeroWidthChar) {
-      const newContent = this.#selectSpan!.textContent!;
-      this.#selectSpan!.textContent = zeroWidthChar;
-      const newElement = this.editor.state.createTextElement(
-        new Delta().insert(newContent)
-      );
-      changeset.forceUpdate = true;
-      changeset.insertChildrenAfter(
-        this.props!.parent as DataBaseElement,
-        [newElement],
-        this.props
-      );
-      changeset.setCursorState(
-        CursorState.collapse(newElement.id, newContent.length)
-      );
-    }
-  }
-
-  override findTextOffsetInBlock(): number {
-    return 0;
-  }
-}
